Feature 4

+++ src/main/java/ru/mail/polis/service/ivlev/ThreadController.java
@@ -1,6 +1,8 @@
 package ru.mail.polis.service.ivlev;
 
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
+import one.nio.http.HttpClient;
+import one.nio.http.HttpException;
 import one.nio.http.HttpServer;
 import one.nio.http.HttpServerConfig;
 import one.nio.http.HttpSession;
@@ -9,15 +11,20 @@ import one.nio.http.Path;
 import one.nio.http.Request;
 import one.nio.http.RequestMethod;
 import one.nio.http.Response;
+import one.nio.net.ConnectionString;
+import one.nio.pool.PoolException;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.mail.polis.dao.DAO;
+import ru.mail.polis.dao.Topology;
 import ru.mail.polis.service.Service;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.ExecutorService;
@@ -30,6 +37,11 @@ public class ThreadController extends HttpServer implements Service {
     private final DAO dao;
     @NotNull
     private final ExecutorService executor;
+    @NotNull
+    private final Topology<String> topology;
+    @NotNull
+    private final Map<String, HttpClient> clients;
+
     private static final Logger logger = LoggerFactory.getLogger(ThreadController.class);
     private static final String RESPONSE_ERROR_LOG_MESSAGE = "Fail send response: ";
 
@@ -45,11 +57,25 @@ public class ThreadController extends HttpServer implements Service {
             final HttpServerConfig config,
             @NotNull final DAO dao,
             final int workersCount,
-            final int queueSize) throws IOException {
+            final int queueSize,
+            @NotNull final Topology<String> topology) throws IOException {
         super(config, dao);
         assert workersCount > 0;
         assert queueSize > 0;
         this.dao = dao;
+        this.topology = topology;
+        this.clients = new HashMap<>();
+        for (final String node : topology.getAllNodes()) {
+            logger.debug(node);
+            if (topology.equalsUrl(node)) {
+                continue;
+            }
+            final HttpClient client = new HttpClient(new ConnectionString(node + "?timeout=1000"));
+            if (clients.put(node, client) != null) {
+                throw new IllegalArgumentException("Duplicate node!");
+            }
+            logger.debug("NODE {}, CLIENT {}", node, client.name());
+        }
         executor = new ThreadPoolExecutor(
                 workersCount, queueSize,
                 0L, TimeUnit.MILLISECONDS,
@@ -63,7 +89,7 @@ public class ThreadController extends HttpServer implements Service {
     }
 
     /**
-     *  End - point status.
+     * End - point status.
      *
      * @param session - http session
      */
@@ -81,28 +107,40 @@ public class ThreadController extends HttpServer implements Service {
     /**
      * End-point get.
      *
-     * @param id - id
+     * @param id      - id
      * @param session - http session
      */
     @Path("/v0/entity")
     @RequestMethod(Request.METHOD_GET)
-    public void get(@Param(value = "id", required = true) final String id, final HttpSession session) {
+    public void get(
+            @Param(value = "id", required = true) final String id,
+            final HttpSession session,
+            final Request request) {
         executor.execute(() -> {
-            try {
-                if (id.isEmpty()) {
-                    session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
-                } else {
-                    session.sendResponse(new Response(Response.OK, toByteArray(dao.get(toByteBuffer(id)))));
-                }
-            } catch (NoSuchElementException e) {
+            final String node = topology.getNodeByKey(id);
+            if (topology.equalsUrl(node)) {
                 try {
-                    session.sendResponse(new Response(Response.NOT_FOUND, Response.EMPTY));
-                } catch (IOException ioException) {
-                    logger.error(RESPONSE_ERROR_LOG_MESSAGE, ioException);
+                    if (id.isEmpty()) {
+                        session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
+                    } else {
+                        session.sendResponse(new Response(Response.OK, toByteArray(dao.get(toByteBuffer(id)))));
+                    }
+                } catch (NoSuchElementException e) {
+                    try {
+                        session.sendResponse(new Response(Response.NOT_FOUND, Response.EMPTY));
+                    } catch (IOException ioException) {
+                        logger.error(RESPONSE_ERROR_LOG_MESSAGE, ioException);
+                    }
+                } catch (IOException e) {
+                    try {
+                        session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    } catch (IOException ioException) {
+                        logger.error(RESPONSE_ERROR_LOG_MESSAGE, ioException);
+                    }
                 }
-            } catch (IOException e) {
+            } else {
                 try {
-                    session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    session.sendResponse(proxy(node, request));
                 } catch (IOException ioException) {
                     logger.error(RESPONSE_ERROR_LOG_MESSAGE, ioException);
                 }
@@ -113,7 +151,7 @@ public class ThreadController extends HttpServer implements Service {
     /**
      * End-point put.
      *
-     * @param id - id
+     * @param id      - id
      * @param session - http session
      * @param request - http request
      */
@@ -124,16 +162,25 @@ public class ThreadController extends HttpServer implements Service {
             final HttpSession session,
             final Request request) {
         executor.execute(() -> {
-            try {
-                if (id.isEmpty()) {
-                    session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
-                } else {
-                    dao.upsert(toByteBuffer(id), ByteBuffer.wrap(request.getBody()));
-                    session.sendResponse(new Response(Response.CREATED, Response.EMPTY));
-                }
-            } catch (IOException e) {
+            final String node = topology.getNodeByKey(id);
+            if (topology.equalsUrl(node)) {
                 try {
-                    session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    if (id.isEmpty()) {
+                        session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
+                    } else {
+                        dao.upsert(toByteBuffer(id), ByteBuffer.wrap(request.getBody()));
+                        session.sendResponse(new Response(Response.CREATED, Response.EMPTY));
+                    }
+                } catch (IOException e) {
+                    try {
+                        session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    } catch (IOException ioException) {
+                        logger.error(RESPONSE_ERROR_LOG_MESSAGE, ioException);
+                    }
+                }
+            } else {
+                try {
+                    session.sendResponse(proxy(node, request));
                 } catch (IOException ioException) {
                     logger.error(RESPONSE_ERROR_LOG_MESSAGE, ioException);
                 }
@@ -144,23 +191,35 @@ public class ThreadController extends HttpServer implements Service {
     /**
      * End-point delete.
      *
-     * @param id - id
+     * @param id      - id
      * @param session - http session
      */
     @Path("/v0/entity")
     @RequestMethod(Request.METHOD_DELETE)
-    public void delete(@Param(value = "id", required = true) final String id, final HttpSession session) {
+    public void delete(
+            @Param(value = "id", required = true) final String id,
+            final HttpSession session,
+            final Request request) {
         executor.execute(() -> {
-            try {
-                if (id.isEmpty()) {
-                    session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
-                } else {
-                    dao.remove(toByteBuffer(id));
-                    session.sendResponse(new Response(Response.ACCEPTED, Response.EMPTY));
-                }
-            } catch (IOException e) {
+            final String node = topology.getNodeByKey(id);
+            if (topology.equalsUrl(node)) {
                 try {
-                    session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    if (id.isEmpty()) {
+                        session.sendResponse(new Response(Response.BAD_REQUEST, Response.EMPTY));
+                    } else {
+                        dao.remove(toByteBuffer(id));
+                        session.sendResponse(new Response(Response.ACCEPTED, Response.EMPTY));
+                    }
+                } catch (IOException e) {
+                    try {
+                        session.sendResponse(new Response(Response.INTERNAL_ERROR, Response.EMPTY));
+                    } catch (IOException ioException) {
+                        logger.error(RESPONSE_ERROR_LOG_MESSAGE, ioException);
+                    }
+                }
+            } else {
+                try {
+                    session.sendResponse(proxy(node, request));
                 } catch (IOException ioException) {
                     logger.error(RESPONSE_ERROR_LOG_MESSAGE, ioException);
                 }
@@ -188,6 +247,9 @@ public class ThreadController extends HttpServer implements Service {
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
         }
+        for (final HttpClient client : clients.values()) {
+            client.close();
+        }
     }
 
     private static ByteBuffer toByteBuffer(@NotNull final String id) {
@@ -201,5 +263,17 @@ public class ThreadController extends HttpServer implements Service {
         final byte[] result = new byte[byteBuffer.remaining()];
         byteBuffer.get(result);
         return result;
+    }
+
+    @NotNull
+    private Response proxy(
+            @NotNull final String node,
+            @NotNull final Request request) {
+        try {
+            return clients.get(node).invoke(request);
+        } catch (IOException | InterruptedException | PoolException | HttpException e) {
+            logger.error("Proxy unavailable: ", e);
+            return new Response(Response.INTERNAL_ERROR, Response.EMPTY);
+        }
     }
 }
